requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
appdirs
attrs
audioop-lts
black
blinker
bs4
cachetools
certifi
cffi
charset-normalizer
click
cryptography
cssselect
discord.py
fake-useragent
flake8
Flask
frozenlist
google-auth
google-genai
h11
httpcore
httpx
idna
importlib_metadata
isort
itsdangerous
Jinja2
lxml
MarkupSafe
mccabe
multidict
mypy
mypy_extensions
packaging
parse
pathspec
pip3-autoremove
platformdirs
propcache
pyasn1
pyasn1_modules
pycodestyle
pycparser
pydantic
pydantic_core
pyee
pyflakes
pyppeteer
pyquery
python-dateutil
requests
requests-html
rsa
setuptools
six
sniffio
tqdm
types-cachetools
types-cffi
types-colorama
types-docutils
types-pyasn1
types-Pygments
types-pyOpenSSL
types-python-dateutil
types-requests
types-setuptools
typing-inspection
typing_extensions
urllib3
w3lib
websockets
Werkzeug
yarl
zipp
lxml_html_clean
